Upload Document _ Raw Algorithms

+++ TM/tm_stm32f4_ds1307.h
@@ -151,13 +151,13 @@ typedef enum {
  * @brief  Structure for date/time
  */
 typedef struct {
-	uint8_t seconds; /*!< Seconds parameter, from 00 to 59 */
-	uint8_t minutes; /*!< Minutes parameter, from 00 to 59 */
-	uint8_t hours;   /*!< Hours parameter, 24Hour mode, 00 to 23 */
-	uint8_t day;     /*!< Day in a week, from 1 to 7 */
-	uint8_t date;    /*!< Date in a month, 1 to 31 */
-	uint8_t month;   /*!< Month in a year, 1 to 12 */
-	uint8_t year;    /*!< Year parameter, 00 to 99, 00 is 2000 and 99 is 2099 */
+	int seconds; /*!< Seconds parameter, from 00 to 59 */
+	int minutes; /*!< Minutes parameter, from 00 to 59 */
+	int hours;   /*!< Hours parameter, 24Hour mode, 00 to 23 */
+	int day;     /*!< Day in a week, from 1 to 7 */
+	int date;    /*!< Date in a month, 1 to 31 */
+	int month;   /*!< Month in a year, 1 to 12 */
+	int year;    /*!< Year parameter, 00 to 99, 00 is 2000 and 99 is 2099 */
 } TM_DS1307_Time_t;
 
 /**
